refactor: move loader packages
