+++ pkg/loaders/resource/load.go
