+++ pkg/loaders/resource/load_test.go
